requirements.txt
streamlit
transformers
gtts
PyPDF2
speechrecognition
pydub
pyaudio
langchain
langchain-community
langchain-core
langchain-groq
langchain-huggingface
langchain-text-splitters
faiss-cpu
torch
pypdf
streamlit-audiorecorder==0.0.3